Angular/thesis-registration/src/app/guards/auth.guard.ts
import { Injectable } from '@angular/core';
import {
  CanActivate,
  ActivatedRouteSnapshot,
  RouterStateSnapshot,
  Router,
} from '@angular/router';

@Injectable({
  providedIn: 'root',
})
export class AuthGuard implements CanActivate {
  constructor(private router: Router) {}

  canActivate(
    next: ActivatedRouteSnapshot,
    state: RouterStateSnapshot
  ): boolean {
    const token = localStorage.getItem('token');
    if (token) {
      return true; // Ha van token, engedélyezzük az útvonalat
    } else {
      this.router.navigate(['/login']);
      return false; // Ha nincs token, visszairányítás a bejelentkezésre
    }
  }
}
